changed all ratings to integer Long type
double values were causing parse errors

+++ app/src/main/java/com/a0xffffffff/dinesum/FirebaseManager.java
@@ -369,7 +369,7 @@ public class FirebaseManager {
      */
     public static User parseUserJson(DataSnapshot userSnapshot) {
         String userID = (String) userSnapshot.child("userID").getValue();
-        double userPoints = ((Long) userSnapshot.child("points").getValue()).doubleValue();
+        Long userPoints = ((Long) userSnapshot.child("points").getValue());
 
         User newUser = new User();
         newUser.setUserID(userID);
